requirements.txt
cffi==1.15.0
pika==1.2.1
pycparser==2.21
rabbitmq==0.2.0
python-dotenv 